+++ gtsfm/multi_view_optimizer.py
@@ -1,0 +1,148 @@
+"""Optimizer which performs averaging and bundle adjustment on all images in
+the scene.
+
+Authors: Ayush Baid, John Lambert
+"""
+import logging
+import sys
+from typing import Any, Dict, List, Optional, Tuple
+
+import dask
+import matplotlib
+
+matplotlib.use("Agg")
+import networkx as nx
+from dask.delayed import Delayed
+from gtsam import Cal3Bundler, PinholeCameraCal3Bundler, Pose3, Rot3, Unit3
+
+import gtsfm.utils.serialization  # import needed to register serialization fns
+from gtsfm.averaging.rotation.rotation_averaging_base import (
+    RotationAveragingBase,
+)
+from gtsfm.averaging.translation.translation_averaging_base import (
+    TranslationAveragingBase,
+)
+from gtsfm.bundle.bundle_adjustment import BundleAdjustmentOptimizer
+from gtsfm.data_association.data_assoc import DataAssociation
+
+# configure loggers to avoid DEBUG level stdout messages
+logging.basicConfig(stream=sys.stdout, level=logging.INFO)
+
+mpl_logger = logging.getLogger("matplotlib")
+mpl_logger.setLevel(logging.WARNING)
+
+pil_logger = logging.getLogger("PIL")
+pil_logger.setLevel(logging.INFO)
+
+
+class MultiViewOptimizer:
+    def __init__(
+        self,
+        rot_avg_module: RotationAveragingBase,
+        trans_avg_module: TranslationAveragingBase,
+        config: Any,
+    ):
+        self.rot_avg_module = rot_avg_module
+        self.trans_avg_module = trans_avg_module
+        self.data_association_module = DataAssociation(
+            config.reproj_error_thresh,
+            config.min_track_len,
+            config.triangulation_mode,
+            config.num_ransac_hypotheses,
+        )
+        self.ba_optimizer = BundleAdjustmentOptimizer()
+
+    def create_computation_graph(
+        self,
+        num_images: int,
+        keypoints_graph: List[Delayed],
+        i2Ri1_graph: Dict[Tuple[int, int], Delayed],
+        i2Ui1_graph: Dict[Tuple[int, int], Delayed],
+        v_corr_idxs_graph: Dict[Tuple[int, int], Delayed],
+        intrinsics_graph: List[Delayed],
+    ) -> Tuple[Delayed, Delayed]:
+        # prune the graph to a single connected component.
+        pruned_graph = dask.delayed(self.select_largest_connected_component)(
+            i2Ri1_graph, i2Ui1_graph
+        )
+
+        pruned_i2Ri1_graph = pruned_graph[0]
+        pruned_i2Ui1_graph = pruned_graph[1]
+
+        wRi_graph = self.rot_avg_module.create_computation_graph(
+            num_images, pruned_i2Ri1_graph
+        )
+
+        wti_graph = self.trans_avg_module.create_computation_graph(
+            num_images, pruned_i2Ui1_graph, wRi_graph
+        )
+
+        init_cameras_graph = dask.delayed(self.init_cameras)(
+            wRi_graph, wti_graph, intrinsics_graph
+        )
+
+        ba_input_graph = self.data_association_module.create_computation_graph(
+            init_cameras_graph, v_corr_idxs_graph, keypoints_graph
+        )
+
+        ba_result_graph = self.ba_optimizer.create_computation_graph(
+            ba_input_graph
+        )
+
+        return ba_input_graph, ba_result_graph
+
+    @classmethod
+    def select_largest_connected_component(
+        cls,
+        rotations: Dict[Tuple[int, int], Optional[Rot3]],
+        unit_translations: Dict[Tuple[int, int], Optional[Unit3]],
+    ) -> Tuple[Dict[Tuple[int, int], Rot3], Dict[Tuple[int, int], Unit3]]:
+        """Process the graph of image indices with Rot3s/Unit3s defining edges,
+        and select the largest connected component."""
+
+        input_edges = [k for (k, v) in rotations.items() if v is not None]
+
+        # create a graph from all edges which have an essential matrix
+        result_graph = nx.Graph()
+        result_graph.add_edges_from(input_edges)
+
+        # get the largest connected components
+        largest_cc = max(nx.connected_components(result_graph), key=len)
+        result_subgraph = result_graph.subgraph(largest_cc).copy()
+
+        # get the remaining edges and construct the dictionary back
+        pruned_edges = list(result_subgraph.edges())
+
+        # as the edges are non-directional, they might have flipped and should
+        # be corrected
+        selected_edges = []
+        for i1, i2 in pruned_edges:
+            if (i1, i2) in rotations:
+                selected_edges.append((i1, i2))
+            else:
+                selected_edges.append((i2, i1))
+
+        # return the subset of original input
+        return (
+            {k: rotations[k] for k in selected_edges},
+            {k: unit_translations[k] for k in selected_edges},
+        )
+
+    @classmethod
+    def init_cameras(
+        cls,
+        wRi_list: List[Optional[Rot3]],
+        wti_list: List[Optional[Unit3]],
+        intrinsics_list: List[Cal3Bundler],
+    ) -> Dict[int, PinholeCameraCal3Bundler]:
+        """Generate camera from valid rotations and unit-translations."""
+
+        cameras = {}
+
+        for idx, (wRi, wti) in enumerate(zip(wRi_list, wti_list)):
+            if wRi is not None and wti is not None:
+                cameras[idx] = PinholeCameraCal3Bundler(
+                    Pose3(wRi, wti), intrinsics_list[idx]
+                )
+
+        return cameras
