+++ gtsfm/scene_optimizer.py
@@ -6,7 +6,7 @@ import logging
 import os
 import sys
 from types import SimpleNamespace
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, List, Optional, Tuple
 
 import dask
 import gtsam
@@ -14,7 +14,6 @@ import matplotlib
 
 matplotlib.use("Agg")
 import matplotlib.pyplot as plt
-import networkx as nx
 from dask.delayed import Delayed
 from dask.distributed import Client, LocalCluster, performance_report
 from gtsam import (
@@ -26,8 +25,8 @@ from gtsam import (
     Unit3,
 )
 
-import gtsfm.utils.io as io_utils
 import gtsfm.utils.geometry_comparisons as comp_utils
+import gtsfm.utils.io as io_utils
 import gtsfm.utils.serialization  # import needed to register serialization fns
 import gtsfm.utils.viz as viz_utils
 from gtsfm.averaging.rotation.rotation_averaging_base import (
@@ -40,12 +39,9 @@ from gtsfm.averaging.translation.averaging_1dsfm import (
 from gtsfm.averaging.translation.translation_averaging_base import (
     TranslationAveragingBase,
 )
-from gtsfm.bundle.bundle_adjustment import BundleAdjustmentOptimizer
 from gtsfm.common.sfm_result import SfmResult
-from gtsfm.data_association.data_assoc import (
-    DataAssociation,
-    TriangulationParam,
-)
+from gtsfm.data_association.data_assoc import TriangulationParam
+from gtsfm.feature_extractor import FeatureExtractor
 from gtsfm.frontend.detector_descriptor.detector_descriptor_base import (
     DetectorDescriptorBase,
 )
@@ -55,6 +51,8 @@ from gtsfm.frontend.matcher.twoway_matcher import TwoWayMatcher
 from gtsfm.frontend.verifier.ransac import Ransac
 from gtsfm.frontend.verifier.verifier_base import VerifierBase
 from gtsfm.loader.folder_loader import FolderLoader
+from gtsfm.multi_view_optimizer import MultiViewOptimizer
+from gtsfm.two_view_estimator import TwoViewEstimator
 
 # configure loggers to avoid DEBUG level stdout messages
 logging.basicConfig(stream=sys.stdout, level=logging.INFO)
@@ -64,175 +62,6 @@ mpl_logger.setLevel(logging.WARNING)
 
 pil_logger = logging.getLogger("PIL")
 pil_logger.setLevel(logging.INFO)
-
-
-class FeatureExtractor:
-    """Wrapper for running detection and description on each image."""
-
-    def __init__(self, detector_descriptor: DetectorDescriptorBase):
-        self.detector_descriptor = detector_descriptor
-
-    def create_computation_graph(
-        self, image_graph: Delayed
-    ) -> Tuple[Delayed, Delayed]:
-        """ Given an image, create detection and descriptor generation tasks """
-        return self.detector_descriptor.create_computation_graph(image_graph)
-
-
-class TwoViewEstimator:
-    """Wrapper for running two-view relative pose estimation on image pairs in
-    the dataset."""
-
-    def __init__(self, matcher: MatcherBase, verifier: VerifierBase):
-        self.matcher = matcher
-        self.verifier = verifier
-
-    def create_computation_graph(
-        self,
-        keypoints_i1_graph: Delayed,
-        keypoints_i2_graph: Delayed,
-        descriptors_i1_graph: Delayed,
-        descriptors_i2_graph: Delayed,
-        camera_intrinsics_i1_graph: Delayed,
-        camera_intrinsics_i2_graph: Delayed,
-        exact_intrinsics: bool = True,
-    ) -> Tuple[Delayed, Delayed, Delayed]:
-        """Create delayed tasks for matching and verification."""
-
-        # graph for matching to obtain putative correspondences
-        corr_idxs_graph = self.matcher.create_computation_graph(
-            descriptors_i1_graph, descriptors_i2_graph
-        )
-
-        # verification on putative correspondences to obtain relative pose
-        # and verified correspondences
-        (
-            i2Ri1_graph,
-            i2Ui1_graph,
-            v_corr_idxs_graph,
-        ) = self.verifier.create_computation_graph(
-            keypoints_i1_graph,
-            keypoints_i2_graph,
-            corr_idxs_graph,
-            camera_intrinsics_i1_graph,
-            camera_intrinsics_i2_graph,
-            exact_intrinsics,
-        )
-
-        return i2Ri1_graph, i2Ui1_graph, v_corr_idxs_graph
-
-
-class MultiViewOptimizer:
-    def __init__(
-        self,
-        rot_avg_module: RotationAveragingBase,
-        trans_avg_module: TranslationAveragingBase,
-        config: Any,
-    ):
-        self.rot_avg_module = rot_avg_module
-        self.trans_avg_module = trans_avg_module
-        self.data_association_module = DataAssociation(
-            config.reproj_error_thresh,
-            config.min_track_len,
-            config.triangulation_mode,
-            config.num_ransac_hypotheses,
-        )
-        self.ba_optimizer = BundleAdjustmentOptimizer()
-
-    def create_computation_graph(
-        self,
-        num_images: int,
-        keypoints_graph: List[Delayed],
-        i2Ri1_graph: Dict[Tuple[int, int], Delayed],
-        i2Ui1_graph: Dict[Tuple[int, int], Delayed],
-        v_corr_idxs_graph: Dict[Tuple[int, int], Delayed],
-        intrinsics_graph: List[Delayed],
-    ) -> Tuple[Delayed, Delayed]:
-        # prune the graph to a single connected component.
-        pruned_graph = dask.delayed(self.select_largest_connected_component)(
-            i2Ri1_graph, i2Ui1_graph
-        )
-
-        pruned_i2Ri1_graph = pruned_graph[0]
-        pruned_i2Ui1_graph = pruned_graph[1]
-
-        wRi_graph = self.rot_avg_module.create_computation_graph(
-            num_images, pruned_i2Ri1_graph
-        )
-
-        wti_graph = self.trans_avg_module.create_computation_graph(
-            num_images, pruned_i2Ui1_graph, wRi_graph
-        )
-
-        init_cameras_graph = dask.delayed(self.init_cameras)(
-            wRi_graph, wti_graph, intrinsics_graph
-        )
-
-        ba_input_graph = self.data_association_module.create_computation_graph(
-            init_cameras_graph, v_corr_idxs_graph, keypoints_graph
-        )
-
-        ba_result_graph = self.ba_optimizer.create_computation_graph(
-            ba_input_graph
-        )
-
-        return ba_input_graph, ba_result_graph
-
-    @classmethod
-    def select_largest_connected_component(
-        cls,
-        rotations: Dict[Tuple[int, int], Optional[Rot3]],
-        unit_translations: Dict[Tuple[int, int], Optional[Unit3]],
-    ) -> Tuple[Dict[Tuple[int, int], Rot3], Dict[Tuple[int, int], Unit3]]:
-        """Process the graph of image indices with Rot3s/Unit3s defining edges,
-        and select the largest connected component."""
-
-        input_edges = [k for (k, v) in rotations.items() if v is not None]
-
-        # create a graph from all edges which have an essential matrix
-        result_graph = nx.Graph()
-        result_graph.add_edges_from(input_edges)
-
-        # get the largest connected components
-        largest_cc = max(nx.connected_components(result_graph), key=len)
-        result_subgraph = result_graph.subgraph(largest_cc).copy()
-
-        # get the remaining edges and construct the dictionary back
-        pruned_edges = list(result_subgraph.edges())
-
-        # as the edges are non-directional, they might have flipped and should
-        # be corrected
-        selected_edges = []
-        for i1, i2 in pruned_edges:
-            if (i1, i2) in rotations:
-                selected_edges.append((i1, i2))
-            else:
-                selected_edges.append((i2, i1))
-
-        # return the subset of original input
-        return (
-            {k: rotations[k] for k in selected_edges},
-            {k: unit_translations[k] for k in selected_edges},
-        )
-
-    @classmethod
-    def init_cameras(
-        cls,
-        wRi_list: List[Optional[Rot3]],
-        wti_list: List[Optional[Unit3]],
-        intrinsics_list: List[Cal3Bundler],
-    ) -> Dict[int, PinholeCameraCal3Bundler]:
-        """Generate camera from valid rotations and unit-translations."""
-
-        cameras = {}
-
-        for idx, (wRi, wti) in enumerate(zip(wRi_list, wti_list)):
-            if wRi is not None and wti is not None:
-                cameras[idx] = PinholeCameraCal3Bundler(
-                    Pose3(wRi, wti), intrinsics_list[idx]
-                )
-
-        return cameras
 
 
 class SceneOptimizer:
